server/recommender/requirements.txt
pandas==1.2.4
psycopg2-binary
nltk==3.6.2
numpy==1.20.2
scikit_learn==0.24.2
